feat: Answer delete, comment, vote 시 token 인증 구현

+++ server/routes/answer.js
@@ -20,7 +20,7 @@ router.post('/:questionId', authenticateToken, async (req, res) => {
     }
     console.log(question.userId.toString(), userIdFromToken);
     if (question.userId.toString() === userIdFromToken) {
-      res.status(401).json('You cannot answer your own question!');
+      res.status(401).json('Access denied');
     }
     const newAnswer = new Answer({
       questionId: req.params.questionId,
@@ -145,6 +145,9 @@ router.put('/:id', authenticateToken, async (req, res) => {
   const userIdFromToken = req.user.id;
   try {
     const answer = await Answer.findById(answerId);
+    if (!answer) {
+      return res.status(404).json('Answer not found!');
+    }
     if (answer.userId.toString() === userIdFromToken) {
       // 수정 사항에 questionId, title 있으면 무시
       delete req.body.questionId;
@@ -165,7 +168,7 @@ router.put('/:id', authenticateToken, async (req, res) => {
         res.status(500).json(err);
       }
     } else {
-      res.status(401).json('You can update only your Answer!');
+      res.status(401).json('Access denied');
     }
   } catch (err) {
     res.status(500).json(err);
@@ -173,19 +176,22 @@ router.put('/:id', authenticateToken, async (req, res) => {
 });
 
 // DELETE
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', authenticateToken, async (req, res) => {
+  const answerId = req.params.id;
+  const userIdFromToken = req.user.id;
   try {
-    const answer = await Answer.findById(req.params.id);
-
+    const answer = await Answer.findById(answerId);
     if (!answer) {
-      res.status(404).json('Answer not found!');
+      return res.status(404).json('Answer not found!');
     }
-
+    if (answer.userId.toString() !== userIdFromToken) {
+      return res.status(401).json('Access denied');
+    }
     try {
       const question = await Question.findById(answer.questionId);
 
-      await Answer.findByIdAndDelete(req.params.id);
-      await Vote.deleteMany({ answerId: req.params.id });
+      await Answer.findByIdAndDelete(answerId);
+      await Vote.deleteMany({ answerId: answerId });
       question.answers -= 1;
       await question.save();
       res.status(200).json('Answer has been deleted');
@@ -200,27 +206,23 @@ router.delete('/:id', async (req, res) => {
 // UPDATE ETC
 
 // Comment
-router.put('/:id/comment', async (req, res) => {
+router.put('/:id/comment', authenticateToken, async (req, res) => {
   const answerId = req.params.id;
-  console.log(answerId);
+  const userIdFromToken = req.user.id;
   try {
     const answer = await Answer.findById(answerId);
-    const userId = req.body.userId;
-    const user = await User.findById(userId);
-
     if (!answer) {
-      res.status(404).json('Answer not found!');
-    }
-    if (!user) {
-      res.status(404).json('User not found!');
+      return res.status(404).json('Answer not found!');
     }
 
     const newComment = new Comment({
       answerId: answerId,
       content: req.body.content,
-      userId: userId,
+      userId: userIdFromToken,
     });
     const savedComment = await newComment.save();
+    answer.comments += 1;
+    await answer.save();
     res.status(200).json(savedComment);
   } catch (err) {
     res.status(500).json(err);
@@ -228,33 +230,37 @@ router.put('/:id/comment', async (req, res) => {
 });
 
 // VOTE
-router.put('/:id/vote', async (req, res) => {
+router.put('/:id/vote', authenticateToken, async (req, res) => {
   const answerId = req.params.id;
-  const userId = req.body.userId;
-
+  const userIdFromToken = req.user.id;
   try {
     const answer = await Answer.findById(answerId);
-
+    let isVoted = false;
     if (!answer) {
-      res.status(404).json('Answer not found!');
+      return res.status(404).json('Answer not found!');
+    }
+    if (answer.userId.toString() === userIdFromToken) {
+      return res.status(401).json('You cannot vote your answer');
     }
     const existingVote = await Vote.findOne({
       answerId,
-      userId: userId,
+      userId: userIdFromToken,
     });
 
     if (!existingVote) {
       await Vote.create({
         answerId,
-        userId: userId,
+        userId: userIdFromToken,
       });
       answer.votes += 1;
+      isVoted = true;
     } else {
       await Vote.deleteOne({ _id: existingVote._id });
       answer.votes -= 1;
+      isVoted = false;
     }
     await answer.save();
-    res.status(200).json(answer);
+    res.status(200).json({ message: 'Vote has been updated', isVoted });
   } catch (err) {
     res.status(500).json(err);
   }
